EMAR-231: changing  long --> int where it was missed by the last merge
git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C364

+++ WebAPI/API/C_Development/C_DevTrunk/Emar.Core/Options/Repository/IOptionRepository.cs
@@ -2,6 +2,6 @@
 {
     public interface IOptionRepository
     {
-        public string GetOption(long siteId, string optionName);
+        public string GetOption(int siteId, string optionName);
     }
 }
